Working Linux scons build.
Various C headers were needed in some of the files. Also, for some reason, Tiny2D_StdFile can't see the File::OpenMode enum. I'm a C guy, so I'll leave that scope issue to someone better versed in such things.

There were a few lines in Tiny2D_OpenGLProcedures.h that needed to be commented. I believe this is an issue specific to my machine, but it might be more general. It looks like that code needs to detect which extensions are available at runtime anyway, so I didn't mess with it too much.